requirements.txt
Flask==2.0.3
openai==0.27.0
transformers==4.28.0
requests==2.28.1
SpeechRecognition==3.10.0
sqlite3
google-auth==2.20.0
google-api-python-client==2.82.0
python-dotenv==1.0.0
textblob==0.17.1
